pyproject.toml
[tool.poetry]
name = "research-assistant"
version = "0.1.0"
description = "Agentic research assistant powered by Claude and web search"
authors = ["Eirini"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
anthropic = "^0.75.0"
streamlit = "^1.52.1"
pyyaml = "^6.0.3"
duckduckgo-search = "^8.1.1"
requests = "^2.32.5"
beautifulsoup4 = "^4.14.3"
trafilatura = "^2.0.0"
python-dotenv = "^1.2.1"


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
black = "^25.12.0"
ruff = "^0.14.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
